// src/apis/models/Amount.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Amount {
  /**
   * Numerical value of the transaction
   */
  'amountValue': number;
  /**
   * Currency of the transaction
   */
  'amountCurrency': string;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'amountValue',
      baseName: 'amountValue',
      type: 'number',
      format: '',
    },
    {
      name: 'amountCurrency',
      baseName: 'amountCurrency',
      type: 'string',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return Amount.attributeTypeMap;
  }

  public constructor() {}
}